fix: Avatar change bug
Desc:
- after updates introduced in commit 17ffe2b31ac32fc8299a3cb825d0dfc20888def2 , avatar changing delete default-avatar.svg for all users, and create broken url to avatar (avatar doesn't exist)

Solution:
- add to UploadAvatarForm request.POST parametr to fix not updating bug (avatar doesn't updates after saving)
- add checking before avatar clean if this avatar isn't default
- in JS after successful avatar changing correct Profile rendering so now Profile Component updates properly

+++ work_training/fifa_league/views.py
@@ -363,15 +363,16 @@ class UserViewSet(viewsets.ViewSet):
             return HttpResponseForbidden()
 
         old_filename = request.user.profile.avatar.name
-
-        form = UserAvatarUploadForm(data=request.FILES,
+        form = UserAvatarUploadForm(request.POST, request.FILES,
                                     instance=request.user.profile)
         if not form.is_valid():
             response = {'is_valid': False,
                         'message': 'Invalid data!'}
             return JsonResponse(response, status=400)
 
-        request.user.profile.avatar.storage.delete(old_filename)
+        if old_filename != '../media/static/fifa_league' \
+                           '/user/default-avatar.svg':
+            request.user.profile.avatar.storage.delete(old_filename)
 
         profile = form.save()
         response = {'is_valid': True,
